music/requirements.txt
# Music Bot Requirements
discord.py>=2.3.0
toml>=0.10.2
python-dotenv>=1.0.0
yt-dlp>=2023.3.4
PyNaCl>=1.5.0
